test: 에러 메시지 출력 확인 테스트 추가
중복된 이름 존재할경우 예외 처리 테스트
게임 실행 횟수가 문장령리 경우 예외 처리 테스트
